SNAP-215: Properly install the snaplock.properties/conf files.
Fixed the handling of the max_connections parameter. It is now clamped between 5 and 1000. Note that the system default these days is 128.
Fixed the default in snapcommunicator to use -1 and thus end up with the default as defined in the snap_communicator.h in the library.
Added a loop in the remove_entering() function to clean any f_still_entering keys that have timed out.
Heavily cleaned up all the .conf file with properly variable definitions, always say what the default value is, etc.

+++ snapwebsites/src/snapdbproxy/snapdbproxy.cpp
@@ -298,7 +298,7 @@ snapdbproxy::snapdbproxy(int argc, char * argv[])
         if(!max_connections.isEmpty())
         {
             bool ok;
-            f_max_pending_connections = max_connections.toLong(&ok);
+            f_max_pending_connections = max_connections.toInt(&ok);
             if(!ok)
             {
                 SNAP_LOG_FATAL("invalid max_pending_connections, a valid number was expected instead of \"")(max_connections)("\".");
